Fix typos, refactor cpp_check and allow checker error messages to propogate

+++ main.py
@@ -4,8 +4,9 @@ from src.stress_tester import StressTester
 
 def main():
     st = StressTester()
-    base_data_path = Path('data').resolve()
-    (base_data_path / 'good').unlink()
+    base_data_path = Path('scripts').resolve()
+    if (base_data_path / 'good').is_file():
+        (base_data_path / 'good').unlink()
     st.set_checker_file(base_data_path / 'default_checker.py')
     st.set_bad_file(base_data_path / 'bad.cpp')
     st.set_gen_file(base_data_path / 'generator.py')
